Add public domain singlebeam bathymetry data for the Caribbean (#2)
This dataset is a compilation of several singlebeam bathymetry surveys
downloaded from the NOAA NCEI. The data are originally in MGD77 format and
include a header with metadata on each survey. The original data file has
around 2 million data points and was compressed with LZMA to save space and
make it possible to upload it to this GitHub repository.
There are some very interesting geodynamic features in this dataset and also
good interpolation examples, since the islands create gaps that many
interpolation methods will struggle to fill. The final compressed CSV with the
data is under 8 Mb.
License: Public domain
Original source: https://ngdc.noaa.gov/mgg/geodas/trackline.html