Add a British Columbia LiDAR point cloud (#7)
This is a section of the LidarBC project (Open Government Licence -
British Columbia), cropped, converted to WGS84, and saved in a
compressed CSV for easier access. The points are ground reflections.
The data are reasonably big (~800k points) and being islands causes some
interpolations to fail when trying to interpolate between the islands.