Add GPS velocity data from the Alps (#8)
This is a compilation of 3D GPS velocities for the Alps by Sánchez et al.
(2018). All velocity components and even the position have error estimates,
which is very useful and rare to find in a lot of datasets. The data is less
neat than the California one we used in Verde but it's good enough, has all of
the uncertainties, a clear citation, and an open license.