Standardize formatting and file names (#12)
Always use "location-type" for the file and folder names to be
consistent. Includes the information from individual folder README.md
files in the main README instead to make it more accessible. Add a
preview figure to each folder to make it easier to inspect without
having to open the notebooks. Clarify the license information.